requirements.txt
discord
python-dotenv
web3
pyyaml